src/libs/python/requirements.txt
pycrypto==2.6.1
pyserial==3.4
termcolor==1.1.0
fastecdsa==2.0.0
cmd2==0.8.0